backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
openai
google-generativeai
python-dotenv